move deprecated definitions
